+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/crypto/model/rest/KeysUploadResponse.kt
@@ -24,7 +24,8 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 internal data class KeysUploadResponse(
         /**
-         * The count per algorithm as returned by the home server: a map (algorithm to count).
+         * Required. For each key algorithm, the number of unclaimed one-time keys
+         * of that type currently held on the server for this device.
          */
         @Json(name = "one_time_key_counts")
         val oneTimeKeyCounts: Map<String, Int>? = null